ADMIN: Wrote system and HTTP tests for admin user remove

See merge request COMP1531/20T1/H13B-THE_CHUNTS!233

+++ src/admin.py
@@ -18,7 +18,13 @@ def admin_userpermission_change(token, u_id, permission_id):
 		Empty Dictionary
 
 	"""
+
+    if None in {token, u_id, permission_id}:
+        raise InputError(description='Insufficient parameters')
+
     token_payload = decode_token(token)
+    u_id = int(u_id)
+    permission_id = int(permission_id)
 
     if u_id not in helpers.get_all_u_id():
         raise InputError(description='u_id does not refer to a valid user')
@@ -46,6 +52,10 @@ def admin_user_remove(token, u_id):
 		Empty Dictionary
 
 	"""
+
+    if None in {token, u_id}:
+        raise InputError(description='Insufficient parameters')
+
     token_payload = decode_token(token)
 
     if u_id not in helpers.get_all_u_id():
@@ -53,6 +63,12 @@ def admin_user_remove(token, u_id):
 
     if not helpers.is_owner(token_payload['u_id']):
         raise AccessError(description='The authorised user is not an owner')
+
+    if token_payload['u_id'] == u_id and len(helpers.get_owners()) == 1:
+        raise InputError(
+            description=
+            'You must assign another user to be an admin before removing yourself'
+        )
 
     helpers.delete_user(u_id)
